fix errors, use core/transaction.Msg rename
