+++ core/transaction/transaction.go
@@ -5,7 +5,7 @@ import (
 )
 
 type (
-	Type     = gogoproto.Message
+	Msg      = gogoproto.Message
 	Identity = []byte
 )
 
@@ -21,7 +21,7 @@ type Tx interface {
 	// Hash returns the unique identifier for the Tx.
 	Hash() [32]byte // TODO evaluate if 32 bytes is the right size & benchmark overhead of hashing instead of using identifier
 	// GetMessages returns the list of state transitions of the Tx.
-	GetMessages() ([]Type, error)
+	GetMessages() ([]Msg, error)
 	// GetSenders returns the tx state transition sender.
 	GetSenders() ([]Identity, error) // TODO reduce this to a single identity if accepted
 	// GetGasLimit returns the gas limit of the tx. Must return math.MaxUint64 for infinite gas
