+++ server/v2/appmanager/appmanager.go
@@ -122,7 +122,7 @@ func (a AppManager[T]) Simulate(ctx context.Context, tx T) (appmanager.TxResult,
 
 // Query queries the application at the provided version.
 // CONTRACT: Version must always be provided, if 0, get latest
-func (a AppManager[T]) Query(ctx context.Context, version uint64, request transaction.Type) (transaction.Type, error) {
+func (a AppManager[T]) Query(ctx context.Context, version uint64, request transaction.Msg) (transaction.Msg, error) {
 	// if version is provided attempt to do a height query.
 	if version != 0 {
 		queryState, err := a.db.StateAt(version)
@@ -146,7 +146,7 @@ func (a AppManager[T]) Query(ctx context.Context, version uint64, request transa
 func (a AppManager[T]) QueryWithState(
 	ctx context.Context,
 	state corestore.ReaderMap,
-	request transaction.Type,
-) (transaction.Type, error) {
+	request transaction.Msg,
+) (transaction.Msg, error) {
 	return a.stf.Query(ctx, state, a.config.QueryGasLimit, request)
 }
