+++ server/v2/stf/export_test.go
@@ -11,4 +11,3 @@ func GetExecutionContext(ctx context.Context) *executionContext {
 	}
 	return executionCtx
 }
-
