+++ server/v2/stf/stf_router.go
@@ -8,8 +8,6 @@ import (
 	appmodulev2 "cosmossdk.io/core/appmodule/v2"
 	gogoproto "github.com/cosmos/gogoproto/proto"
 	"google.golang.org/protobuf/proto"
-
-	appmodulev2 "cosmossdk.io/core/appmodule/v2"
 )
 
 var ErrNoHandler = errors.New("no handler")
